KAD-41: jdoc test-api & add more abstraction to `EntityBuilder`
