+++ lib/kadai-test-api/src/main/java/io/kadai/testapi/CleanKadaiContext.java
@@ -23,6 +23,27 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
+/**
+ * Annotation allowing to declare the cleansing of the KadaiContext in
+ * {@linkplain org.junit.jupiter.api.Nested @Nested} Test-Classes.
+ *
+ * <p>It makes the annotated class reuse the surrounding classes'
+ * {@linkplain javax.sql.DataSource DataSource} but generates a new schema.
+ *
+ * <p>Usage may look like:
+ * <pre>
+ *   {@code
+ *     @ExtendWith(KadaiInitializationExtension.class)
+ *     class MyTestClass {
+ *
+ *       @Nested
+ *       @CleanKadaiContext
+ *       @TestInstance(Lifecycle.PER_CLASS)
+ *       class MyNestedTestClass {}
+ *     }
+ *   }
+ * </pre>
+ */
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.TYPE)
 public @interface CleanKadaiContext {}
