+++ lib/kadai-test-api/src/main/java/io/kadai/testapi/KadaiConfigurationModifier.java
@@ -20,7 +20,35 @@ package io.kadai.testapi;
 
 import io.kadai.KadaiConfiguration;
 
+/**
+ * Interface specifying operations allowing to modify the {@link KadaiConfiguration} of an
+ * integration test extended with
+ * {@linkplain
+ * io.kadai.testapi.extensions.KadaiInitializationExtension @KadaiInitializationExtension}.
+ *
+ * <p>Usage may look like:
+ * <pre>
+ *   {@code
+ *     @ExtendWith(KadaiInitializationExtension.class)
+ *     class MyTestClass implements KadaiEngineConfigurationModifier {
+ *       @Override
+ *       public KadaiConfiguration.Builder modify(KadaiConfiguration.Builder builder) {
+ *         return builder.foo();
+ *       }
+ *
+ *       @Test
+ *       void myTest() {}
+ *     }
+ *   }
+ *  * </pre>
+ */
 public interface KadaiConfigurationModifier {
 
+  /**
+   * Modifies the {@link KadaiConfiguration.Builder}.
+   *
+   * @param builder the builder to modify
+   * @return the modified builder
+   */
   KadaiConfiguration.Builder modify(KadaiConfiguration.Builder builder);
 }
