+++ lib/kadai-test-api/src/main/java/io/kadai/testapi/builder/Builder.java
@@ -1,0 +1,16 @@
+package io.kadai.testapi.builder;
+
+/**
+ * Interface specifying how to build entities.
+ *
+ * @param <EntityT> the type of entity to build and store
+ */
+public interface Builder<EntityT> {
+
+  /**
+   * Builds the {@linkplain EntityT entity} for this builder.
+   *
+   * @return the built entity
+   */
+  EntityT build();
+}
