+++ lib/kadai-test-api/src/main/java/io/kadai/testapi/builder/EntityBuilder.java
@@ -19,14 +19,36 @@
 package io.kadai.testapi.builder;
 
 import io.kadai.common.api.security.UserPrincipal;
+import io.kadai.user.api.models.User;
 import java.security.PrivilegedActionException;
 import java.security.PrivilegedExceptionAction;
 import javax.security.auth.Subject;
 
-public interface EntityBuilder<EntityT, ServiceT> {
+/**
+ * Interface specifying how to build and store entities in the database for integration tests.
+ *
+ * @param <EntityT> the type of entity to build and store
+ * @param <ServiceT> the type of the service for storing the entity
+ */
+public interface EntityBuilder<EntityT, ServiceT> extends Builder<EntityT> {
 
+  /**
+   * Builds the {@linkplain EntityT entity} for this builder and stores it in the database.
+   *
+   * @param service the {@linkplain ServiceT service} storing the built entity
+   * @return the stored entity
+   * @throws Exception if building or storing the entity fails
+   */
   EntityT buildAndStore(ServiceT service) throws Exception;
 
+  /**
+   * Builds the {@linkplain EntityT entity} for this builder and stores it in the database.
+   *
+   * @param service the {@linkplain ServiceT service} storing the built entity
+   * @param userId the {@linkplain User#getId() id} of the user to store the built entity as
+   * @return the stored entity
+   * @throws Exception if building or storing the entity fails
+   */
   default EntityT buildAndStore(ServiceT service, String userId) throws Exception {
     return execAsUser(userId, () -> buildAndStore(service));
   }
@@ -37,19 +59,5 @@ public interface EntityBuilder<EntityT, ServiceT> {
     subject.getPrincipals().add(new UserPrincipal(userId));
 
     return Subject.doAs(subject, runnable);
-  }
-
-  interface SummaryEntityBuilder<SummaryEntityT, EntityT extends SummaryEntityT, ServiceT>
-      extends EntityBuilder<EntityT, ServiceT> {
-    SummaryEntityT entityToSummary(EntityT entity);
-
-    default SummaryEntityT buildAndStoreAsSummary(ServiceT service) throws Exception {
-      return entityToSummary(buildAndStore(service));
-    }
-
-    default SummaryEntityT buildAndStoreAsSummary(ServiceT service, String userId)
-        throws Exception {
-      return entityToSummary(buildAndStore(service, userId));
-    }
   }
 }
