+++ lib/kadai-test-api/src/main/java/io/kadai/testapi/builder/TaskCommentBuilder.java
@@ -77,4 +77,9 @@ public class TaskCommentBuilder implements EntityBuilder<TaskComment, TaskServic
       testTaskComment.setId(null);
     }
   }
+
+  @Override
+  public TaskComment build() {
+    return testTaskComment.copy();
+  }
 }
