+++ lib/kadai-test-api/src/main/java/io/kadai/testapi/extensions/TestContainerExtension.java
@@ -39,6 +39,9 @@ import org.junit.jupiter.api.extension.InvocationInterceptor;
 import org.junit.jupiter.api.extension.ReflectiveInvocationContext;
 import org.testcontainers.containers.JdbcDatabaseContainer;
 
+/**
+ * JUnit-Extension for creating containerized databases.
+ */
 public class TestContainerExtension implements InvocationInterceptor {
 
   public static final String STORE_DATA_SOURCE = "datasource";
